Fix gwtc (can only hash file plaintext up to 4 GiB)

+++ src/peergos/shared/crypto/hash/ScryptJS.java
@@ -33,7 +33,7 @@ public class ScryptJS implements Hasher {
 
     @Override
     public CompletableFuture<Multihash> hash(AsyncReader stream, long length) {
-        return scriptJS.streamSha256(stream, length)
+        return scriptJS.streamSha256(stream, (int) length)
                 .thenApply(h -> new Multihash(Multihash.Type.sha2_256, h));
     }
 }
